% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\name{ace}
\alias{ace}
\title{rollapply to a data frame}
\usage{
ace(directory, input, output, sampleNum, g2)
}
\arguments{
\item{directory}{the working directory}

\item{input}{the input file name}

\item{output}{the ouput file name}

\item{sampleNum}{the length of the time series}

\item{g2}{the L2 regulazation strength}
}
\description{
rollapply to a data frame
}
