% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_product_correction.R
\name{average_product_correction}
\alias{average_product_correction}
\title{average product correction for normal matrix}
\usage{
average_product_correction(norm_mat)
}
\arguments{
\item{norm_mat}{the normal matrix}
}
\description{
average product correction for normal matrix
}
