% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/AverageDegree.R
\name{AverageDegree}
\alias{AverageDegree}
\title{Function used to calculate the averge degree of a graph}
\usage{
AverageDegree(GRAPH)
}
\arguments{
\item{GRAPH}{A igraph object}
}
\value{
A numeric value
}
\description{
Function used to calculate the averge degree of a graph
}

