% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/AverageTrappingTime.R
\name{averageTrappingTime}
\alias{averageTrappingTime}
\title{Function used to calculate the averge trapping time for a graph}
\usage{
averageTrappingTime(GRAPH)
}
\arguments{
\item{GRAPH}{A igraph object}
}
\value{
A numeric value
}
\description{
Function used to calculate the averge trapping time for a graph
}

