% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycenter.r
\name{barycenter}
\alias{barycenter}
\title{calculate the barycenter of the trade network}
\usage{
barycenter(net)
}
\arguments{
\item{net}{a tbl_graph object}
}
\value{
a tibble with import and export latitude and lontitude
}
\description{
calculate the barycenter of the trade network
}
\references{
https://www.mckinsey.com/featured-insights/urbanization/urban-world-cities-and-the-rise-of-the-consuming-class
}
