% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/basicGraphTopology.R
\name{basicGraphTopology}
\alias{basicGraphTopology}
\title{Function used to calculate the basic network topology}
\usage{
basicGraphTopology(GRAPH)
}
\arguments{
\item{GRAPH}{a igraph object}
}
\value{
a vector with six elements
}
\description{
Function used to calculate the basic network topology
}

