% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump_chart.r
\name{bump_chart}
\alias{bump_chart}
\title{bump chart or ranking chart for the uncomtrade data}
\usage{
bump_chart(
  net_list,
  Y,
  rank_limit = 10,
  rank_range = c(30, 1),
  size_range = c(3, 10)
)
}
\arguments{
\item{net_list}{}

\item{rank_limit}{the largest ranking number of countries}

\item{y}{the ranking end year, for example 2017}

\item{rank_name}{the variable name use in dataframe d to perform ranking}

\item{trade_flow}{the ranking is for export or import}
}
\value{
a ggplot object
}
\description{
bump chart or ranking chart for the uncomtrade data
}
