% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadingFailure.R
\name{cascadingFailure}
\alias{cascadingFailure}
\title{This program is trying to modeling the cascading failure}
\usage{
cascadingFailure(GRAPH1, GRAPH2, attack = "random", p = 0.7, q = NULL,
  plot = FALSE, vids = NULL)
}
\arguments{
\item{GRAPH1}{the first layer of inter dependent network}

\item{GRAPH2}{the scond layer of inter dependent network}

\item{attack}{the type of attack, random, target}

\item{p}{1-p is the fraction of nodes that the initial attack start}

\item{q}{the coupling strength between two layers}

\item{plot}{whether to plot the network or not}
}
\value{
a vector with length two give the node numbers of the giant component
}
\description{
This program is trying to modeling the cascading failure
}
