% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{caviar_ADAPTIVE}
\alias{caviar_ADAPTIVE}
\title{adaptive model for CoViaR}
\usage{
caviar_ADAPTIVE(K, THETA, BETA, y, empiricalQuantile, VaR, RowsOfy,
  varPredict)
}
\arguments{
\item{k}{}
}
\value{
a list
}
\description{
adaptive model for CoViaR
}
