% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{caviar_AS}
\alias{caviar_AS}
\title{Asyomptic slope model for CoViaR}
\usage{
caviar_AS(BETA, y, empiricalQuantile, VaR, RowsOfy, varPredict)
}
\value{
a list
}
\description{
Asyomptic slope model for CoViaR
}
