% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caviar.R
\name{caviarOptim}
\alias{caviarOptim}
\title{Function used to calculate the CAViaR}
\usage{
caviarOptim(
  data,
  model = 1,
  pval = 0.01,
  k = 5,
  REP = 5,
  MAXITER = 500,
  predict = F
)
}
\arguments{
\item{data}{a vector of return time series}

\item{model}{choose the model among four different models}

\item{pval}{the significant level}

\item{k}{optinal parameter for ADAPTIVE model}

\item{REP}{number of hessian optimization procedure repeats}

\item{MAXITER}{maximum number of optimization procedure iterations}

\item{predict}{predicted value at risk for the next period (tail(VaR,1))}
}
\value{
a list
}
\description{
Function used to calculate the CAViaR
}
