% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation2net.r
\name{citation2net}
\alias{citation2net}
\title{convert the citation information to co-occurrence network}
\usage{
citation2net(database)
}
\arguments{
\item{database}{A dataframe with authors, institutions, keywords at each row}
}
\value{
an weighted and named igraph object
}
\description{
convert the citation information to co-occurrence network
}
