% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_infomap.r
\name{cluster_infomap_new}
\alias{cluster_infomap_new}
\title{the latest version of infomap algorithm for directed and weighted network}
\usage{
cluster_infomap_new(net)
}
\arguments{
\item{net}{a igraph object}
}
\value{
a vector for the community membership of the vertex, different from the membership class in igraph package
}
\description{
the latest version of infomap algorithm for directed and weighted network
}
