% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community2sankey.r
\name{community2sankey}
\alias{community2sankey}
\title{plot community as sankey diagram}
\usage{
community2sankey(net_list, font_size = 18, scaled = FALSE,
  sinksRight = TRUE, nodePadding = 10, nodeWidth = 30)
}
\arguments{
\item{net_list}{a list of igraph objects}

\item{font_size}{the font size of the plot}

\item{scaled}{whether the edge weight of overlapping nodes between two successive communities should be normalized or not}

\item{sinksRight}{should the plot align to right or not}

\item{nodePadding}{the width height}

\item{nodeWidth}{the width of the nodes}
}
\value{
a list with sankey data and the sankey diagram object
}
\description{
plot community as sankey diagram
}
