% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityCentrality.R
\name{communityCentrality}
\alias{communityCentrality}
\title{community centrality of network}
\usage{
communityCentrality(graph, diag = FALSE)
}
\arguments{
\item{graph}{an igraph object}

\item{diag}{set the diag elements of null model matrix to zero}
}
\value{
the modularity matrix
}
\description{
community centrality of network
}
