% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_network.r
\name{competition_network}
\alias{competition_network}
\title{construct the competition network from the trade network}
\usage{
competition_network(trade_net)
}
\arguments{
\item{trade_net}{a weighted trade network constructed from trade2net() function which use the uncomtrade data}
}
\value{
a list with export and import competition networks
}
\description{
construct the competition network from the trade network
}
