% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict_resolve.r
\name{conflict_resolve}
\alias{conflict_resolve}
\title{resolve the function conflict for one pkg over any other pkgs}
\usage{
conflict_resolve(pkg)
}
\arguments{
\item{pkg}{the name of the package that wants to win}
}
\value{
invisible
}
\description{
resolve the function conflict for one pkg over any other pkgs
}
