% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMap.R
\name{contactMap}
\alias{contactMap}
\title{calculate the contact map from the interaction strength matrix}
\usage{
contactMap(fileName, n)
}
\arguments{
\item{fileName}{the input files with external field and interaction strength}

\item{n}{the system size, skip the first n lines, external field}
}
\value{
a matrix of contact map
}
\description{
calculate the contact map from the interaction strength matrix
}
