% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMat2PMFG.r
\name{corMat2PMFG}
\alias{corMat2PMFG}
\title{Planar Graph from the correlation matrix of financial time series}
\usage{
corMat2PMFG(
  corMat,
  outputFile = NULL,
  descending = TRUE,
  format = c("edgelist", "pajek", "ncol", "lgl", "graphml", "dimacs", "gml", "dot",
    "leda"),
  weighted = TRUE
)
}
\arguments{
\item{corMat}{a correlation matrix}

\item{outputFile}{a string as the name of output file}

\item{descending}{whether arrange the correlation value in descending order or not}

\item{format}{the output file format for graph object}

\item{weighted}{default is TRUE for the weighted one}
}
\value{
a igraph object or edge_list file in current directory
}
\description{
Planar Graph from the correlation matrix of financial time series
}
\references{
A Tool for Filtering Information in Complex Systems, PNAS
}
