% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNames.r
\name{correctNames}
\alias{correctNames}
\title{This function is aiming at converting author names to standard format: family name, Short. Short.(Zhu, Z. W.;Keller, A. Z.), remove blank space}
\usage{
correctNames(x)
}
\arguments{
\item{x}{a string represents a name or a vector}
}
\value{
a string
}
\description{
This function is aiming at converting author names to standard format: family name, Short. Short.(Zhu, Z. W.;Keller, A. Z.), remove blank space
}
