% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.edges_vertices.R
\name{delete.edges_vertices}
\alias{delete.edges_vertices}
\title{Delete edges end with specific nodes}
\usage{
delete.edges_vertices(GRAPH, v, mode = c("out", "in", "all", "total"))
}
\arguments{
\item{GRAPH}{an igraph object}

\item{v}{a vector of vids}

\item{mode}{delete the outgoing or incoming edges or both, ignored for undirecte graphs}
}
\value{
an igraph object
}
\description{
Delete edges end with specific nodes
}
