% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAllPackages.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{Function used to detach all the packages except basics pkgs}
\usage{
detachAllPackages()
}
\description{
Function used to detach all the packages except basics pkgs
}
