% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagModify.R
\name{diagModify}
\alias{diagModify}
\title{change the diag function in order to complain with pipe operator}
\usage{
diagModify(x, diagValue)
}
\arguments{
\item{x}{a matrix}

\item{diagValue}{a numeric value or a vector}
}
\description{
change the diag function in order to complain with pipe operator
}
