% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directSingularity.R
\name{directSingularity}
\alias{directSingularity}
\title{Direct determine the sigularity spectrum of fractal set}
\usage{
directSingularity(x, nVec, sampleNum, thread, qVec)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{sampleNum}{the random sampled number}

\item{thread}{the parallel thread number}

\item{qVec}{the qth order}
}
\value{
the holder exponent and singularity spectrum
}
\description{
Direct determine the sigularity spectrum of fractal set
}
