% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_by_sd.R
\name{discrete_by_sd}
\alias{discrete_by_sd}
\title{discretize the return of stock by standard deviation}
\usage{
discrete_by_sd(x, sd_magnitude = 2)
}
\arguments{
\item{x}{the return time series}

\item{sd_magnitude}{how many sd should use as threshold}
}
\description{
discretize the return of stock by standard deviation
}
