% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_by_volatility.R
\name{discrete_by_volatility}
\alias{discrete_by_volatility}
\title{discretize the volatility of stock by standard deviation}
\usage{
discrete_by_volatility(x, sd_magnitude = 2)
}
\arguments{
\item{x}{the return OR volatility time series}

\item{sd_magnitude}{how many sd should use as threshold}
}
\value{
a vector for the return time series
}
\description{
discretize the volatility of stock by standard deviation
}
