% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econophysics.R
\docType{package}
\name{econophyscis}
\alias{econophyscis}
\title{econophysics: A package for econophysics research.
The econophysics package provides three categories of important functions}
\section{Econophysics functions}{

The econophysics functions ...
}

