% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/effcientFrontierIpop.R
\name{eff.frontier_ipop}
\alias{eff.frontier_ipop}
\title{get the efficient frontier, no matter the covariance matrix is postive semidefinite or not}
\usage{
eff.frontier_ipop(returns, covMat = NULL, short = "yes",
  max.allocation = NULL, risk.premium.up = 20, risk.increment = 0.5,
  thread = 3)
}
\arguments{
\item{covMat}{The predefined covariance matrix}

\item{short}{whether short selling is prohibited or not}

\item{max.allocation}{the maximum allocation of the asset, can control the over conceration}

\item{risk.increment}{the incremental of the risk aversion}

\item{thread}{the multi threads number}

\item{retruns}{The return matrix of all the assets,T*N dimension}

\item{risk.premim}{the maximum risk aversion parameter}
}
\value{
a data.frame include the weight, variance, return and shape ratio
}
\description{
get the efficient frontier, no matter the covariance matrix is postive semidefinite or not
}

