% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenBoundRMT.R
\name{eigenBoundRMT}
\alias{eigenBoundRMT}
\title{calculate the eigenvalue distribution for a random purely random matix}
\usage{
eigenBoundRMT(Q = NULL, L, N, S)
}
\arguments{
\item{L}{the length of the time series}

\item{N}{the number of the time series
return a vector with the lower and upper bound}
}
\description{
calculate the eigenvalue distribution for a random purely random matix
}
