% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenDistRMT.R
\name{eigenDistRMT}
\alias{eigenDistRMT}
\title{calculate the eigenvalue distribution for a random purely random matix}
\usage{
eigenDistRMT(Q = NULL, L, N, S, n)
}
\arguments{
\item{L}{the length of the time series}

\item{N}{the number of the time series}

\item{S}{the standeviation}

\item{b}{the bin length
return a dataframe with the lower and upper bound}
}
\description{
calculate the eigenvalue distribution for a random purely random matix
}
