% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenRev.R
\name{eigenRev}
\alias{eigenRev}
\title{Function used to reverse the order of the eigen vectors}
\usage{
eigenRev(x, symmetric, only.values = FALSE, EISPACK = FALSE)
}
\arguments{
\item{EISPACK}{return a list or a data.frame}
}
\description{
Function used to reverse the order of the eigen vectors
}
