% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementWiseAdd.R
\name{elementWiseAdd}
\alias{elementWiseAdd}
\title{element wise add function for matrix}
\usage{
elementWiseAdd(x)
}
\arguments{
\item{x}{a list of matrices}
}
\value{
a matrix
}
\description{
element wise add function for matrix
}
