% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errbarXY.R
\name{errbarXY}
\alias{errbarXY}
\title{Function used to add the horizonal errbar and vertical errbar}
\usage{
errbarXY(
  x,
  y,
  xDelta = NULL,
  yDelta = NULL,
  col = "red",
  type = "o",
  pch_h = "|",
  pch_v = "-"
)
}
\arguments{
\item{x}{the x coordinate of data}

\item{y}{the y coordinate of data}

\item{xDelta}{the statndrad devation of x which will be used to plot the errbar}

\item{yDelta}{the statndrad devation of y which will be used to plot the errbar}
}
\value{
invisible, a plot
}
\description{
Function used to add the horizonal errbar and vertical errbar
}
