% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESI_INDEX.r
\name{ESI}
\alias{ESI}
\title{the export similarity index for the trade data and will return a ESI matrix if you have many commodities}
\usage{
ESI(
  trade_data,
  commodity_codes,
  Year = TRUE,
  partner_average = TRUE,
  country_names = NULL
)
}
\arguments{
\item{trade_data}{the tibble covnerted by the polish_data function}

\item{commodity_codes}{the codes of the commodities based on which the ESI is calculated}

\item{Year}{split the trade_data by year variable or not, if true the final output will be a list of ESI matrices, otherwise just a single matrix}

\item{partner_average}{if true, it will combine all the partners(importers here) together to calculate the trade value or volume, otherwise the final ESI index will be the average of the ESI for each importer}
}
\value{
a list of ESI matrices if the Year parameter is true, otherwise a single matrix
}
\description{
the export similarity index for the trade data and will return a ESI matrix if you have many commodities
}
\references{
A Measure of `Export Similarity' and Its Possible Uses, Author(s): J. M. Finger and M. E. Kreinin, Source: The Economic Journal, Vol. 89, No. 356 (Dec., 1979), pp. 905-912
}
