% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESI_ab.r
\name{ESI_ab}
\alias{ESI_ab}
\title{the export similarity index for the trade data, only calculate the}
\usage{
ESI_ab(trade_data_tmp, a, b, commodity_codes)
}
\description{
the export similarity index for the trade data, only calculate the
}
