% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedShortfall.R
\name{expectedShortfall}
\alias{expectedShortfall}
\title{use the box constriant to calculate the expected shortfall of a portfolio}
\usage{
expectedShortfall(tickers, returns)
}
\arguments{
\item{tickers}{the names of the stocks}

\item{returns}{the return time series of stocks, log return}
}
\value{
a numeric value
}
\description{
use the box constriant to calculate the expected shortfall of a portfolio
}
