% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_graph.r
\name{extract_graph}
\alias{extract_graph}
\title{Extract a sub-graph according to the ratio of total edge weight}
\usage{
extract_graph(GRAPH, prob = 0.8)
}
\arguments{
\item{GRAPH}{an igraph object}

\item{prob}{the ratio of the total edge weight}
}
\value{
an igraph object
}
\description{
Extract a sub-graph according to the ratio of total edge weight
}
