% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/F2_DFA.R
\name{F2_DFA}
\alias{F2_DFA}
\title{Function used to calculate the square DFA fluctuation}
\usage{
F2_DFA(x, nVec = NULL, sampleNum = NULL, thread = 3, detrendOrder = 3,
  sampleMethod = 1)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{sampleNum}{the random sampled number}

\item{thread}{the parallel threads number}

\item{detrendOrder}{the detrending polynomial order}

\item{sampleMethod}{1 means the determined sample number method, other values mean the nonoverlap method}
}
\value{
A list of the square DFA fluctuation for scale nVec
}
\description{
Function used to calculate the square DFA fluctuation
}

