% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_MFDCCA.R
\name{F_ARq_DCCA}
\alias{F_ARq_DCCA}
\title{ARMA based MFDCCA analysis}
\usage{
F_ARq_DCCA(x, y, nVec = NULL, sampleNum = NULL, qVec = c(-5:5),
  thread = 3, detrendMethod = "AR", sampleMethod = 2,
  lengthRatio = 0.05)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{y}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{sampleNum}{the number of sample}

\item{qVec}{the order of the multifractal}

\item{thread}{the multithread number}

\item{detrendMethod}{the detrending polynomial order}

\item{sampleMethod}{1 means the determined sample number, other values means the nonoverlap method}

\item{lengthRatio}{detremine how long is the largest scale}
}
\value{
the the sqrt DCCA fluctuation
}
\description{
ARMA based MFDCCA analysis
}
