% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_DCCA_seq.R
\name{F_DCCA_seq}
\alias{F_DCCA_seq}
\title{parallel compute the dfa fluctuations for different scale}
\usage{
F_DCCA_seq(x, y, nVec, overlap = 0.9999999, thread = 3, q = 2)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{y}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{overlap}{the overlap of the boxes}

\item{thread}{the multithreads number}

\item{q}{the order of the multifractal}
}
\value{
DCCA fluctuation sequence for different n, a dataframe
}
\description{
parallel compute the dfa fluctuations for different scale
}
