% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/F_DFA.R
\name{F_DFA}
\alias{F_DFA}
\title{Function used to calculate the DFA fluctuation}
\usage{
F_DFA(x, nVec = NULL, sampleNum = NULL, qVec = c(-5:5), thread = 3,
  detrendOrder = 3, sampleMethod = 2)
}
\arguments{
\item{x}{the time series}

\item{nVec}{the detrended length vector}

\item{sampleNum}{the number of random sample}

\item{qVec}{the mutifractal order vector}

\item{thread}{the parallel threads number}

\item{detrendOrder}{the detrending polynomial order}

\item{sampleMethod}{1 means the determined sample number method, other values mean the nonoverlap method}
}
\value{
the DFA fluctuation which enhanced by q^th order, a dataframe
}
\description{
Function used to calculate the DFA fluctuation
}

