% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factorialMoment.R
\name{factorialMoment}
\alias{factorialMoment}
\title{Function to calculate the factor moments}
\usage{
factorialMoment(x)
}
\arguments{
\item{mat1}{the first matrix}

\item{mat2}{the second matrix}
}
\value{
a value between -1 and 1
}
\description{
Function to calculate the factor moments
}

