% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredCorMat.R
\name{filteredCorMat}
\alias{filteredCorMat}
\title{Function used to filter the covariance matrix of a series of return time series}
\usage{
filteredCorMat(corMat, eigenNum = 1)
}
\arguments{
\item{corMat}{the correlation matrix}

\item{eigenNum}{the number of eigevalues be removed}
}
\value{
a martix
}
\description{
Function used to filter the covariance matrix of a series of return time series
}
\references{
Random matrix approach to cross correlations in financial data
}
