% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_cycles.R
\name{find_cycles}
\alias{find_cycles}
\title{Function used to calculate the averge degree of a graph}
\usage{
find_cycles(x)
}
\arguments{
\item{x}{a time series or}
}
\value{
two numeric values, the cycle length of local maxima and minimia
}
\description{
Function used to calculate the averge degree of a graph
}

