% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGiantComponent.R
\name{findGiantComponent}
\alias{findGiantComponent}
\title{Find the giant component}
\usage{
findGiantComponent(GRAPH)
}
\arguments{
\item{GRAPH}{a igraph object}
}
\value{
the size of the giant component
}
\description{
Find the giant component
}
