% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSmallComponent.R
\name{findSmallComponentsVid}
\alias{findSmallComponentsVid}
\title{Find the vertices' ids outside the giant component}
\usage{
findSmallComponentsVid(GRAPH)
}
\arguments{
\item{GRAPH}{an igraph object}
}
\value{
a vector of vertices' ids
}
\description{
Find the vertices' ids outside the giant component
}
