% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finish_nofity.r
\name{finish_notify}
\alias{finish_notify}
\title{this function can use to generate a beep sound after you finish a long task}
\usage{
finish_notify(sound = 1)
}
\arguments{
\item{sound}{numbers to choose different sounds}
}
\value{
invisible
}
\description{
this function can use to generate a beep sound after you finish a long task
}
