% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fq_DCCA.R
\name{Fq_DCCA}
\alias{Fq_DCCA}
\title{Function used to calculate the DCCA power of fluctuation}
\usage{
Fq_DCCA(x, y, nVec = NULL, sampleNum = NULL, qVec = c(-5:5),
  detrendOrder = 3, thread = 3, sampleMethod = 2, lengthRatio = 0.05)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{y}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{sampleNum}{the number of sample}

\item{qVec}{the order of the multifractal}

\item{detrendOrder}{the detrending polynomial order}

\item{thread}{the multithread number}

\item{sampleMethod}{1 means the determined sample number, other values means the nonoverlap method}

\item{lengthRatio}{detremine how long is the largest scale}
}
\value{
the the sqrt DCCA fluctuation
}
\description{
Function used to calculate the DCCA power of fluctuation
}
