% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funcEidtDistance.R
\name{funcEditDistance}
\alias{funcEditDistance}
\title{Function to calculate a simple edit distance}
\usage{
funcEditDistance(GRAPH1, GRAPH2)
}
\arguments{
\item{GRAPH1}{a igraph object}

\item{GRAPH2}{a igraph object}
}
\value{
A numeric value ranging from 0 to 1
}
\description{
Function to calculate a simple edit distance
}

