% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyofNetwork.R
\name{funcEntropyofNetwork}
\alias{funcEntropyofNetwork}
\title{Function used to calculate the entropy of a graph}
\usage{
funcEntropyofNetwork(GRAPH)
}
\arguments{
\item{GRAPH}{a complex network, must be connected}
}
\value{
A numeric value entropy
}
\description{
Function used to calculate the entropy of a graph
}
