% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFromCorMatrixToPMFGEdlist.R
\name{funcGetPlanarFromCorMat}
\alias{funcGetPlanarFromCorMat}
\title{Planar Graph Edgelist}
\usage{
funcGetPlanarFromCorMat(corMat, outputFile = NULL, desending = TRUE)
}
\arguments{
\item{corMat}{a correlation matrix}

\item{outputFile}{a string as the name of outout file}
}
\value{
a edgelist file in current directory
}
\description{
Planar Graph Edgelist
}
