% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/funcGraphEnergy.R
\name{funcGraphEnergy}
\alias{funcGraphEnergy}
\title{Function to calculate the energy of a graph which is the sum of the abs of all the eigenvalue}
\usage{
funcGraphEnergy(GRAPH)
}
\arguments{
\item{GRAPH}{a igraph object}
}
\value{
Return a numeric value
}
\description{
Function to calculate the energy of a graph which is the sum of the abs of all the eigenvalue
}

