% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcHeteroIndex.R
\name{funcHeteroIndex}
\alias{funcHeteroIndex}
\title{Function to eavaluate the heterogeneity of a graph}
\usage{
funcHeteroIndex(GRAPH)
}
\arguments{
\item{GRAPH}{a igraph object}
}
\value{
A numeric value ranges from 0 to 1
}
\description{
Function to eavaluate the heterogeneity of a graph
}
