% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcLapEnergy.R
\name{funcLaplacianEnergy}
\alias{funcLaplacianEnergy}
\title{Function to calculate the laplacian energy of a graph, weighted or unweighted}
\usage{
funcLaplacianEnergy(GRAPH, normalized = FALSE, weights = NULL)
}
\arguments{
\item{GRAPH}{is a igraph object,}

\item{normalized}{means use the normalized energy or not}

\item{weights}{the weights atrribute of a graph object}
}
\value{
A numeric value
}
\description{
Function to calculate the laplacian energy of a graph, weighted or unweighted
}
