% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{funcPlanarTest}
\alias{funcPlanarTest}
\title{Test a graph is planar or not}
\usage{
funcPlanarTest(edgelist, VNum)
}
\arguments{
\item{edgelist}{a edgelist dataframe without edge weights}

\item{VNum}{vertex number of a graph which related to the edgelist}
}
\value{
FALSE OR TRUE
}
\description{
Test a graph is planar or not
}

