% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_nMF.R
\name{h_nMF}
\alias{h_nMF}
\title{naive mean field external field reconstruction for ising model}
\usage{
h_nMF(returns, J_nMF)
}
\arguments{
\item{J_nMF}{a interaction strength J matrix}

\item{retruns}{the return data.frame for stock market OR the configuation matrix for other system}
}
\value{
a vector of external field for every site
}
\description{
naive mean field external field reconstruction for ising model
}
