% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_TAP.R
\name{h_TAP}
\alias{h_TAP}
\title{TAP approximation external field reconstruction for ising model}
\usage{
h_TAP(returns, J_TAP)
}
\arguments{
\item{returns}{the return data.frame for stock market OR the configuation matrix for other system}

\item{J_TAP}{a interaction strength J matrix}
}
\value{
a vector of external field for every site
}
\description{
TAP approximation external field reconstruction for ising model
}
