% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardThresholdNetwork.R
\name{hardThresholdNetwork}
\alias{hardThresholdNetwork}
\title{Construct the hard threshold network form correlation matrix}
\usage{
hardThresholdNetwork(MAT, threshold = 0.1)
}
\description{
Construct the hard threshold network form correlation matrix
}
