% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histDensity.R
\name{histDensity}
\alias{histDensity}
\title{get hist density}
\usage{
histDensity(x, breaks = 100)
}
\arguments{
\item{x}{the dataset}

\item{breaks}{the number of bins}
}
\value{
a data.frame
}
\description{
get hist density
}
