% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/hurstExponent.R
\name{hurstExponent}
\alias{hurstExponent}
\title{Function used to calculate the generalized hurst exponent}
\usage{
hurstExponent(x, nVec = NULL, sampleNum = NULL, thread = 3,
  qVec = c(-5:5), detrendOrder = 3, sampleMethod = 1)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation, the minmum scale should be determined by calculate the autocorrelation length of the time series.}

\item{sampleNum}{the random sampled number}

\item{thread}{the parallel thread number}

\item{qVec}{the qth order}

\item{detrendOrder}{the detrending polynomial order}

\item{sampleMethod}{1 means the determined sample number method, other values mean the nonoverlap method}
}
\value{
the generalized hurst exponent
}
\description{
Function used to calculate the generalized hurst exponent
}

