% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hurstExponentMultiscale.R
\name{hurstExponentMultiscale}
\alias{hurstExponentMultiscale}
\title{Function to cnstrcut the multiscale hurst exponent}
\usage{
hurstExponentMultiscale()
}
\value{
multiscale hurst exponent
}
\description{
Function to cnstrcut the multiscale hurst exponent
}

