% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridMeasure.R
\name{hybirdMeasure}
\alias{hybirdMeasure}
\title{Function calculate hybird measure for centrality of graph}
\usage{
hybirdMeasure(GRAPH_W)
}
\arguments{
\item{GRAPH_W}{an weighted igraph object}
}
\value{
a list or a data.frame with colnames X,Y,X-Y,X+Y, the variable X+Y can be used as compound centrality
}
\description{
Function calculate hybird measure for centrality of graph
}
