% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initializePars.R
\name{initializePars}
\alias{initializePars}
\title{Function used to initialize some system parameters}
\usage{
initializePars()
}
\value{
can not be seen
}
\description{
Function used to initialize some system parameters
}

