% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_ace.R
\name{input_ace}
\alias{input_ace}
\title{convert from configuration to state probability and joint probability}
\usage{
input_ace(x, p_theta = 0.05, thread = 3, fileName)
}
\arguments{
\item{x}{a data.frame for configuration}

\item{p_theta}{the threashold for the probability}

\item{thread}{the multithread number}

\item{fileName}{the outpu file name}
}
\value{
a file
}
\description{
convert from configuration to state probability and joint probability
}
