% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseParticipateRatio.R
\name{inverseParticipateRatio}
\alias{inverseParticipateRatio}
\title{Function used to calculate the inverse participate ratio}
\usage{
inverseParticipateRatio(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
a vector of length equals to the number of the eigenvalue
}
\description{
Function used to calculate the inverse participate ratio
}
