% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isometricTrans}
\alias{isometricTrans}
\title{Function to get the edgelist form a time series}
\usage{
isometricTrans(x, epsilon)
}
\arguments{
\item{x}{a numeric vector convert from a time series}

\item{epsilon}{the threshold}
}
\value{
a edgelist dataframe with weights with 2 columns
}
\description{
Function to get the edgelist form a time series
}
\references{
Geometrical invariability of transformation between a time series and a complex network, PHYSICAL REVIEW E 90, 012804 (2014)
}

