% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_nMF.R
\name{J_nMF}
\alias{J_nMF}
\title{naive mean field copuling strength reconstruction for ising model}
\usage{
J_nMF(returns)
}
\arguments{
\item{returns}{the return data.frame for stock market OR the configuation matrix for other system}
}
\value{
a interaction strength J matrix
}
\description{
naive mean field copuling strength reconstruction for ising model
}
