% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_TAP.R
\name{J_TAP}
\alias{J_TAP}
\title{TAP approximation of copuling strength reconstruction for ising model}
\usage{
J_TAP(returns)
}
\arguments{
\item{returns}{the return data.frame for stock market OR the configuation matrix for other system}
}
\value{
a interaction strength J matrix
}
\description{
TAP approximation of copuling strength reconstruction for ising model
}
