% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_Distance.R
\name{JS_Distance}
\alias{JS_Distance}
\title{Function used to calculate the Jensen-Shannon between two graphs, i.e, the distance between eigenvalue distributions of laplacian matrix of graphs}
\usage{
JS_Distance(GRAPH1, GRAPH2, normalized = TRUE, n = 512, bw = 0.01)
}
\arguments{
\item{GRAPH1}{graph object of igraph, undirected}

\item{GRAPH2}{graph object of igraph , undirected}

\item{normalized}{compute the nornalized laplacian or not}

\item{n}{the number of equally spaced points at which the density is to be estimated}

\item{bw}{the bandwidth of the density function}
}
\value{
a numeric value between 0 and 1
}
\description{
Function used to calculate the Jensen-Shannon between two graphs, i.e, the distance between eigenvalue distributions of laplacian matrix of graphs
}
