% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_ProbDistance.R
\name{JS_ProbDistance}
\alias{JS_ProbDistance}
\title{Function used to calculate the Jensen-Shannon between two discrete probability distribution distribution}
\usage{
JS_ProbDistance(p1, p2)
}
\arguments{
\item{p1}{a discrete probability density function}

\item{p2}{a discrete probability density function}
}
\value{
a numeric value between 0 and 1
}
\description{
Function used to calculate the Jensen-Shannon between two discrete probability distribution distribution
}
