% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_SeriesDistance.R
\name{JS_SeriesDistance}
\alias{JS_SeriesDistance}
\title{Function used to calculate the Jeanson shannon distance for two time sereis}
\usage{
JS_SeriesDistance(x, y, n = 512, bw = 0.01, thread = 3)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{n}{the number of equally spaced points at which the density is to be estimated}

\item{bw}{the bandwidth of the density function}

\item{thread}{the multi threads}
}
\value{
a numeric value between 0 and 1
}
\description{
Function used to calculate the Jeanson shannon distance for two time sereis
}
