% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuramoto.R
\name{kuramoto}
\alias{kuramoto}
\title{simulate the kuramoto model}
\usage{
kuramoto(
  graph = NULL,
  adjMat,
  h = 0.01,
  phase = runif(N, 0, 2 * pi),
  natFreq = rnorm(N),
  thread = 3,
  steps = 1000,
  lambda = 0.1,
  weight = TRUE,
  N = vcount(graph)
)
}
\arguments{
\item{graph}{the network structure, can be weighted}

\item{h}{the stepsize for RK method}

\item{phase}{the initial phase of the oscillators}

\item{natFreq}{the nature frequency of those oscillators}

\item{thread}{the multi threads number}

\item{steps}{the number of steps of the simulation}

\item{lambda}{the coupling strength between the oscillators}

\item{weight}{a logical value, whether use edge weight or not}

\item{N}{the size of the network}
}
\value{
a list with the order parameter and the phase
}
\description{
simulate the kuramoto model
}
