% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendreTransform.R
\name{legendreTransform}
\alias{legendreTransform}
\title{Use the legendre transformation to calculate the lengendre spectrum from the generalized Hurst exponent}
\usage{
legendreTransform(h_q, qVec, D_f = 1)
}
\arguments{
\item{h_q}{the generalized Hurst exponent}

\item{qVec}{the fluctuation order}

\item{D_f}{the dimension of the support, default is 1 for time sereis}
}
\description{
Use the legendre transformation to calculate the lengendre spectrum from the generalized Hurst exponent
}
