% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/localMinMax.R
\name{localMinMax}
\alias{localMinMax}
\title{Function to find the local minimum and maximum of a time series}
\usage{
localMinMax(x, stepT)
}
\arguments{
\item{x}{the time sereis}

\item{stepT}{the moving window to find the local min and max}
}
\value{
a two elements list, the first is the local min time index and the second is local max time index
}
\description{
Function to find the local minimum and maximum of a time series
}

