% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularityMatrix.R
\name{modularityMatrix}
\alias{modularityMatrix}
\title{modularity matrix for netowrk}
\usage{
modularityMatrix(graph, diag = FALSE)
}
\arguments{
\item{graph}{an igraph object}

\item{diag}{set the diag elements of matrix to zero or not}
}
\value{
the modularity matrix
}
\description{
modularity matrix for netowrk
}
