% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2sankey.r
\name{net2sankey}
\alias{net2sankey}
\title{plot the network with sankey diagram, either by country or by continent}
\usage{
net2sankey(net, percent = NULL, fontSize = 25,
  fontFamily = "Times New Roman", nodeWidth = 20, nodePadding = 10,
  sinksRight = FALSE)
}
\arguments{
\item{net}{an igraph or tidygraph object}

\item{fontSize}{the font size of the plot}

\item{fontFamily}{the font type of the plt}

\item{nodeWidth}{the width of the nodes}

\item{nodePadding}{the width of the height}

\item{sinksRight}{whehter the plot will align to the right}

\item{country}{logical value, defalut TRUE, the node of diagram is country or not}
}
\value{
a list with the sankey data for both country and continent and the sankey digram object for the country and continent with different coloring scheme
}
\description{
plot the network with sankey diagram, either by country or by continent
}
