% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network2map.r
\name{network2map}
\alias{network2map}
\title{plot network on world map}
\usage{
network2map(
  GRAPH,
  node_size = NULL,
  node_color = NULL,
  inter_com_size = 0.5,
  inner_com_size = 0.8,
  label_color = NULL,
  label_size = NULL,
  hue = "blue",
  luminosity = "bright",
  inter_com_color = "red",
  inter_com_type = 2,
  random_color = FALSE
)
}
\arguments{
\item{GRAPH}{an igraph object}

\item{node_size}{the predefined node size for the plot, maybe determined inside the func}

\item{node_color}{the predefined node color for the plot, maybe determined inside the func}

\item{inter_com_size}{inter community link size}

\item{label_color}{the color of the node label}

\item{label_size}{the size of the node label}

\item{hue}{the central color of the random generate color code}

\item{luminosity}{the color type, "random", "light", "bright", "dark"}

\item{inter_com_color}{the inter community link colors}

\item{inter_com_type}{the inter community link type}

\item{random_color}{the way to generate color, random or not}
}
\value{
an ggplot object
}
\description{
plot network on world map
}
