% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_backbone.r
\name{network_backbone}
\alias{network_backbone}
\title{extract the network backbone based on the null model proposed by PNAS paper:Extracting the multiscale backbone of complex weighted networks}
\usage{
network_backbone(net, alpha = 0.1, cut_mode = c("or", "and"))
}
\arguments{
\item{net}{a tbl_graph object from tidygraph pkg}

\item{alpha}{the significant level}

\item{cut_mode}{parameter valid for the directed network, "or" means the alpha_ij value for}
}
\value{
a igraph object
}
\description{
extract the network backbone based on the null model proposed by PNAS paper:Extracting the multiscale backbone of complex weighted networks
}
