% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_node.r
\name{node_ranking}
\alias{node_ranking}
\title{ranking the node importance for the trade network}
\usage{
node_ranking(GRAPH, YEAR)
}
\arguments{
\item{GRAPH}{a weighted network}

\item{YEAR}{the year info}
}
\value{
a tibble
}
\description{
ranking the node importance for the trade network
}
