% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NodeTrappingTime.R
\name{nodeTrappingTime}
\alias{nodeTrappingTime}
\title{Function used to calculate the averge trapping time for a node}
\usage{
nodeTrappingTime(GRAPH, vids = V(GRAPH))
}
\arguments{
\item{GRAPH}{A igraph object}

\item{vids}{the vertex ids}
}
\value{
A numeric value
}
\description{
Function used to calculate the averge trapping time for a node
}

