% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalES.R
\name{optimalES}
\alias{optimalES}
\title{use the DEoptim algorithm to calculate the optimal ES by using the box constriant}
\usage{
optimalES(stockNames, returns, wMin = 0, wMax = 1, searchSize = 50000,
  itermax = 200, min_sum = 0.99, max_sum = 1.01, p = 0.95)
}
\arguments{
\item{stockNames}{the ticker of stocks}

\item{returns}{the return matrix of a bunch of stocks}

\item{wMin}{the minimum weight of each stock}

\item{wMax}{the maximum weight of each stock}

\item{searchSize}{the search steps of DEoptim algorithm}

\item{min_sum}{the minimum of the weight sum}

\item{max_sum}{the maximum of the weight sum}

\item{p}{the confidence level of the ES}

\item{intermax}{the parameter for DEoptim.control function}
}
\value{
A numeric value
}
\description{
use the DEoptim algorithm to calculate the optimal ES by using the box constriant
}
