% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dfa.R
\name{plot.dfa}
\alias{plot.dfa}
\title{plot the fluctuation function for DFA}
\usage{
\method{plot}{dfa}(x, xlab = "s", ylab = "fq", ylim = NULL, tck = 0.02)
}
\arguments{
\item{x}{the time series}

\item{xlab}{the x label}

\item{ylab}{the y label}

\item{ylim}{the y coordinate range}

\item{tck}{the plot tck value for the parameter option}
}
\value{
invisible
}
\description{
plot the fluctuation function for DFA
}
