% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerMapping.R
\name{powerMapping}
\alias{powerMapping}
\title{power mapping the correlation matrix}
\usage{
powerMapping(corMat, q = 1.5)
}
\arguments{
\item{corMat}{the original correlation matrix}

\item{q}{the order of the power mapping}
}
\value{
a matrix after power mapping
}
\description{
power mapping the correlation matrix
}
