% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probaMeasure.R
\name{probaMeasure}
\alias{probaMeasure}
\title{Calculate the probability measure of a fractal set}
\usage{
probaMeasure(x, nVec, sampleNum, thread)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{sampleNum}{the random sampled number}

\item{thread}{the parallel thread number}
}
\value{
the unnormalized probability measure, a data.frame
}
\description{
Calculate the probability measure of a fractal set
}
