% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomAttackSingleNetwork.r
\name{randomAttackSingleNetwork}
\alias{randomAttackSingleNetwork}
\title{Random attack for single network}
\usage{
randomAttackSingleNetwork(GRAPH, p = 0.1, vids = NULL,
  attack = c("random", "target", "localized", "motif"), mode = c("out",
  "in", "all", "total"), decreasing = TRUE, normalized = TRUE)
}
\arguments{
\item{GRAPH}{a igraph object}

\item{p}{the fraction of remaining nodes}

\item{vids}{the selected initial attack nodes in the target and localized attack}

\item{attack}{the type of attack for random, target and localized and motif}

\item{mode}{whether to delete the outgoing incoming or both edges for selected nodes}

\item{decreasing}{sort degree from large to small or not}

\item{normalized}{normalize the number of the giant component}

\item{normalized}{the giant component size}
}
\value{
a number of the giant component node
}
\description{
Random attack for single network
}
