% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomShuffle.R
\name{randomShuffle}
\alias{randomShuffle}
\title{Function used to random shuffle some time series}
\usage{
randomShuffle(x, n)
}
\arguments{
\item{x}{the time series you want to shuffle}

\item{n}{how many times you want to shuffle}
}
\value{
a vector
}
\description{
Function used to random shuffle some time series
}

