% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomTargetAttackSingleNetwork.r
\name{randomTargetAttackSingleNetwork}
\alias{randomTargetAttackSingleNetwork}
\title{attack a network with different ways}
\usage{
randomTargetAttackSingleNetwork(GRAPH, l = 10, vids = NULL,
  attack = c("random", "target", "localized", "motif"),
  decreasing = TRUE, normalized = TRUE)
}
\arguments{
\item{GRAPH}{igraph object}

\item{l}{the largets attack steps}

\item{vids}{the vertex ids}

\item{attack}{the type of attack for random, target and localized and motif}

\item{decreasing}{sort degree from large to small or not}

\item{normalized}{normalize the number of the giant component}

\item{normalized}{normalized the giant component size}
}
\value{
a number of the giant component node
}
\description{
attack a network with different ways
}
