% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thresholdNetworkStatistics.R
\name{rankBasedThresholdNetworkStatiscs}
\alias{rankBasedThresholdNetworkStatiscs}
\title{Function used to construct the rank based threashold network}
\usage{
rankBasedThresholdNetworkStatiscs(MAT, mix = FALSE, d = 2)
}
\arguments{
\item{MAT}{the original correlation matrix}

\item{mix}{whether use the mixed rank or not}

\item{d}{the d_th largest elements should be used}
}
\value{
a dataframe
}
\description{
Function used to construct the rank based threashold network
}

