% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rho_DPCCA.R
\name{rho_DPCCA}
\alias{rho_DPCCA}
\title{Function used to calculate the DPCCA partial coefficient}
\usage{
rho_DPCCA(corMat)
}
\arguments{
\item{corMat}{the DCCA correlation matrix}
}
\value{
the partial correlation matrix
}
\description{
Function used to calculate the DPCCA partial coefficient
}

