% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rungeKutta.R
\name{RK}
\alias{RK}
\title{runge kutta method for kuramoto model}
\usage{
RK(theta_i, theta_j, h, couple_vec, natFreq)
}
\arguments{
\item{theta_i}{the phase for the focal oscillator}

\item{theta_j}{the phase vector for the neighbor of i}

\item{h}{the stepsize of the RK method}

\item{couple_vec}{the couping vector for i and its neighbors}

\item{natFreq}{the nature frequency for the oscillator}
}
\value{
the phase of the focal oscillator at next time
}
\description{
runge kutta method for kuramoto model
}
