% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollapply.data.frame.R
\name{rollapply.data.frame}
\alias{rollapply.data.frame}
\title{rollapply to a data frame}
\usage{
\method{rollapply}{data.frame}(x, margin = 1, FUN = NULL, ...,
  thread = 3, windowSize = NULL, stepSize = NULL,
  returnAsTbl = TRUE)
}
\arguments{
\item{x}{the data frame}

\item{margin}{whic dimension you want to apply}

\item{FUN}{the function}

\item{...}{other parameters for FUN}

\item{thread}{the multithreds number}

\item{windowSize}{the size of the rolling window}

\item{stepSize}{the moving window size}
}
\value{
a list
}
\description{
rollapply to a data frame
}
