% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV_Mat.R
\name{RV_Mat}
\alias{RV_Mat}
\title{Function to calculate the modified RV coefficient, i.e, the correlation coefficient between matrices}
\usage{
RV_Mat(mat1, mat2)
}
\arguments{
\item{mat1}{the first matrix}

\item{mat2}{the second matrix}
}
\value{
a value between -1 and 1
}
\description{
Function to calculate the modified RV coefficient, i.e, the correlation coefficient between matrices
}
