% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSave.R
\name{safeSave}
\alias{safeSave}
\title{safe save for objects}
\usage{
safeSave(..., file = stop("'file' must be specified"),
  overwrite = FALSE, save.fun = save.image)
}
\arguments{
\item{...}{the objects}

\item{file}{the file name to store the objects}

\item{overwrite}{overwrite the existed file or not}

\item{the}{function used to save the objects}
}
\value{
invisible
}
\description{
safe save for objects
}
