% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPortfoiloCommunity.r
\name{selectPortfolioCommunity}
\alias{selectPortfolioCommunity}
\title{Select stocks based the community detection info,inside community}
\usage{
selectPortfolioCommunity(
  GRAPH,
  mem = NULL,
  mem_algo = "cluster_infomap_new",
  stock_id = NULL,
  portfolioSize = 4,
  type = c("inside", "inter"),
  central_rule = c("sum", "mean")
)
}
\arguments{
\item{GRAPH}{the correlation-based graph}

\item{mem}{the community membership of the GRAPH, default NULL}

\item{mem_algo}{the algorithm to perform the community detection}

\item{stock_id}{the names of the stocks, if the graph is unamed, will be used to name the nodes}

\item{portfolioSize}{the size of the portfolio, ie., the stocks you want to pick from the graph}

\item{type}{"inside" means the stocks are all picked from one community with very strong connection, "inter" means the stocks are all picked from different communities}

\item{central_rule}{the rule use to compare different community by the hybird centrality measure when choose one stocks over another}
}
\value{
a vector with stock names or id
}
\description{
Select stocks based the community detection info,inside community
}
