% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPortfoiloInterCommunity.r
\name{selectPortfolioInterCommunity}
\alias{selectPortfolioInterCommunity}
\title{Select sotcks based the community detection info}
\usage{
selectPortfolioInterCommunity(GRAPH, mem, stock_id, portfolioSize = 4)
}
\arguments{
\item{GRAPH}{the correlation-based graph}

\item{mem}{the community membership of te GRAPH}

\item{stock_id}{the names of the stocks}

\item{portfolioSize}{the size of the portfolio}
}
\description{
Select sotcks based the community detection info
}
