% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/showMemoryUse.R
\name{showMemoryUse}
\alias{showMemoryUse}
\title{Function used to show the memory useage in a readable format}
\usage{
showMemoryUse(sort = "size", decreasing = TRUE, limit = 20)
}
\arguments{
\item{sort}{sort by which attribute}

\item{decreasing}{sort by decreasing order or not}

\item{limit}{the maximum objects listed}
}
\description{
Function used to show the memory useage in a readable format
}

