% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPkgContents.r
\name{showPackageContents}
\alias{showPackageContents}
\title{Show the functions and objects in a pkg}
\usage{
showPackageContents(packageName)
}
\arguments{
\item{packageName}{a character}
}
\value{
a list
}
\description{
Show the functions and objects in a pkg
}
