% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.r
\name{spillover}
\alias{spillover}
\title{calculate the spillover relationship between two time series}
\usage{
spillover(r1, r2, VaR1, VaR2, M, BETA)
}
\arguments{
\item{r1}{a time series}

\item{r2}{a time series}

\item{VaR1}{the VaR of r1}

\item{VaR2}{the VaR of r2}

\item{M}{the lag parameter use to calculate the spillover}

\item{BETA}{the significant threshold for the causality test}
}
\value{
a vector with two element, 1 or 0
}
\description{
calculate the spillover relationship between two time series
}
