% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRCA.r
\name{SRCA}
\alias{SRCA}
\title{the RCA and Normalized RCA}
\usage{
SRCA(trade_data, commodity_code, Year = TRUE, country_names = NULL)
}
\arguments{
\item{trade_data}{the trade dataset}

\item{commodity_code}{a vector of commodity codes SITC V3}

\item{Year}{whether split the trade_data set by year or aggregate the data}

\item{country_names}{the countries that be used to calculate the RCA index}
}
\value{
a named vector of RCA and NRCA
}
\description{
the RCA and Normalized RCA
}
