% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCI_INDEX_V2.r
\name{TCI_v2}
\alias{TCI_v2}
\title{Trade complementarity index by Micbael Michaely}
\usage{
TCI_v2(trade_data, commodity_code, Year = TRUE, country_names = NULL)
}
\arguments{
\item{trade_data}{the polished trade data tibble}

\item{commodity_code}{the commodity code vector}

\item{Year}{logical value, whether the trade_data should be splited by year or not}

\item{country_names}{the name vector of the countries under consideration}
}
\value{
a list of TCI matrices if Year is true, otherwise a single matrix
}
\description{
Trade complementarity index by Micbael Michaely
}
\references{
Trade Preferential Agreements in Latin America An Ex-Ante Assessment, Micbael Michaely
}
