% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_Drawdown.r
\name{tidy_draw_down}
\alias{tidy_draw_down}
\title{the drawdown of a vector accompany to tidyverse}
\usage{
tidy_draw_down(Ra, geometric)
}
\arguments{
\item{Ra}{the time series vector}

\item{geometric}{whether to use geometric or not}
}
\value{

}
\description{
the drawdown of a vector accompany to tidyverse
}
