% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade2net.r
\name{trade_data2net}
\alias{trade_data2net}
\title{convert the uncomtrade data tibble to igraph networks}
\usage{
trade_data2net(trade_data)
}
\arguments{
\item{trade_data}{a tidyverse like tibble generate by the "polish_data" function}
}
\value{
a tidygraph object
}
\description{
convert the uncomtrade data tibble to igraph networks
}
