% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_net_map.r
\name{trade_net_map}
\alias{trade_net_map}
\title{plot the trade network on map and color the country according to the core and periphery}
\usage{
trade_net_map(net, core_node_col = "red",
  periphery_node_col = "orange", mode = c("all", "out", "in"))
}
\arguments{
\item{net}{the trade network}

\item{core_node_col}{the core node color determined by the coreness ranking}

\item{periphery_node_col}{the color of the periphery node}

\item{mode}{the type of the strength}
}
\value{
a ggplot object
}
\description{
plot the trade network on map and color the country according to the core and periphery
}
