% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_clustering_coefficent.r
\name{transitivity_cor_mat}
\alias{transitivity_cor_mat}
\title{The modified clustering coefficient for a complete weighted graph, especially when we treat the correlation matrix as a network}
\usage{
transitivity_cor_mat(corMat, samples = 1000, threads = 14)
}
\arguments{
\item{corMat}{the correlation matrix for returns}
}
\value{
the clustering coefficient vector
}
\description{
The modified clustering coefficient for a complete weighted graph, especially when we treat the correlation matrix as a network
}
