% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplePredictSpectrum.R
\name{triplePredictSpectrum}
\alias{triplePredictSpectrum}
\title{Compute the triplets to evaluate the complexity of the fractal spectrum}
\usage{
triplePredictSpectrum(Alpha, fAlpha, tol = 1e-04)
}
\arguments{
\item{Alpha}{the holder exponent}

\item{fAlpha}{the probability density of the holder exponent}

\item{tol}{the tolerence used to determine the x_min and x_max in the fractal spectrum}
}
\value{
A vector with three elements
}
\description{
Compute the triplets to evaluate the complexity of the fractal spectrum
}
