% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unweightVisibilityGraph}
\alias{unweightVisibilityGraph}
\title{Function to get the edgelist form a time series}
\usage{
unweightVisibilityGraph(x)
}
\arguments{
\item{x}{a numeric vector convert from a time series}
}
\value{
a edgelist dataframe without weights with 2 columns
}
\description{
Function to get the edgelist form a time series
}
