% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibilityGraph.r
\name{visibilityGraph}
\alias{visibilityGraph}
\title{generate the visibility graph from time series}
\usage{
visibilityGraph(x, weighted = FALSE)
}
\arguments{
\item{x}{a time series}

\item{weighted}{a logical, whether create the weighted visibility graph or not}
}
\value{
g a igraph object
}
\description{
generate the visibility graph from time series
}
