% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatilityConnectedness.r
\name{volatilityConnectedness}
\alias{volatilityConnectedness}
\title{calculate the volatility connectedness among financial time series}
\usage{
volatilityConnectedness(x, y)
}
\description{
calculate the volatility connectedness among financial time series
}
