% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightedVisibilityGraph}
\alias{weightedVisibilityGraph}
\title{Function to get the edgelist form a time series}
\usage{
weightedVisibilityGraph(x)
}
\arguments{
\item{x}{a numeric vector convert from a time series}
}
\value{
a edgelist dataframe with weights with 3 columns
}
\description{
Function to get the edgelist form a time series
}
\references{
From time series to complex networks:The visibility graph,4972–4975, PNAS, April 1, 2008, vol. 105, no. 13
}
