% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump_chart.r
\name{bump_chart}
\alias{bump_chart}
\title{bump chart or ranking chart for the uncomtrade data}
\usage{
bump_chart(
  d,
  y,
  rank_limit = 10,
  rank_name,
  trade_flow = c("Export", "Import")
)
}
\arguments{
\item{d}{the ranking dataframe provide by the node_ranking() function}

\item{y}{the ranking end year, for example 2017}

\item{rank_limit}{the largest ranking number of countries}

\item{rank_name}{the variable name use in dataframe d to perform ranking}

\item{trade_flow}{the ranking is for export or import}
}
\value{
a ggplot object
}
\description{
bump chart or ranking chart for the uncomtrade data
}
