% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordDiagramFromNet.r
\name{chordDiagramFromNet}
\alias{chordDiagramFromNet}
\title{plot the chord diagram for the uncomtrade among different continent from a network}
\usage{
chordDiagramFromNet(GRAPH, dat = FALSE, weight_factor = 1)
}
\arguments{
\item{GRAPH}{a tidygraph object}

\item{dat}{Logical value, whether return the trade flow data or not}
}
\value{
a plot
}
\description{
plot the chord diagram for the uncomtrade among different continent from a network
}
