% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_infomap_new.r, R/func_infomap.r
\name{cluster_infomap_new}
\alias{cluster_infomap_new}
\title{the latest version of infomap algorithm}
\usage{
cluster_infomap_new(net)

cluster_infomap_new(net)
}
\arguments{
\item{net}{the igraph object}

\item{net}{a igraph object}
}
\value{
the community objects from the igraph package

a community class, the same as in igraph package
}
\description{
the latest version of infomap algorithm

the latest version of infomap algorithm for directed and weighted network
}
