% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_infomap_new.r
\name{cluster_infomap_new}
\alias{cluster_infomap_new}
\title{the latest version of infomap algorithm}
\usage{
cluster_infomap_new(graph, multilevel = FALSE)
}
\arguments{
\item{graph}{the igraph object}

\item{multilevel}{whether use the multilevel algorithm or not, default is 2 level detection}
}
\value{
the community objects from the igraph package
}
\description{
the latest version of infomap algorithm
}
