% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMap.R
\name{contactMap}
\alias{contactMap}
\title{calculate the contact map from the interaction strength matrix}
\usage{
contactMap(fileName, n)
}
\arguments{
\item{fileName}{the normal matrix}

\item{n}{the system size, skip the first n lines, external field}
}
\description{
calculate the contact map from the interaction strength matrix
}
