% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcFromCorMatrixToPMFGEdlist.R
\name{corMat2PMFG}
\alias{corMat2PMFG}
\title{Planar Graph Edgelist}
\usage{
corMat2PMFG(corMat, outputFile = NULL, desending = TRUE,
  format = c("edgelist", "pajek", "ncol", "lgl", "graphml", "dimacs",
  "gml", "dot", "leda"))
}
\arguments{
\item{corMat}{a correlation matrix}

\item{outputFile}{a string as the name of outout file}

\item{descending}{}
}
\value{
a edgelist file in current directory
}
\description{
Planar Graph Edgelist
}
