% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Function used to create a simple project subdirectories}
\usage{
createProject(projectName)
}
\arguments{
\item{projectName}{the name of the project, relative path or absolute path}
}
\description{
Function used to create a simple project subdirectories
}

