% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.edges_vertices.R
\name{delete.edges_vertices}
\alias{delete.edges_vertices}
\title{Delete edges end with specific nodes}
\usage{
delete.edges_vertices(GRAPH, v)
}
\arguments{
\item{GRAPH}{an igraph object}

\item{v}{a vector of vids}
}
\value{
an igraph object
}
\description{
Delete edges end with specific nodes
}
