% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/efficientFrontier.R
\name{eff.frontier}
\alias{eff.frontier}
\title{Function used to find the efficient frontier by using the modern portofilio theory}
\usage{
eff.frontier(returns, covMat = NULL, short = "no", max.allocation = NULL,
  risk.premium.up = 20, thread = 3)
}
\arguments{
\item{returns}{argument should be a m x n matrix with one column per security}

\item{covMat}{the covariance matrix}

\item{short}{short selling or not}

\item{max.allocation}{is the maximum % allowed for any one security- reduces concentration}

\item{risk.premium.up}{is the upper limit of the risk premium modeled}

\item{thread}{the multithreads argument}
}
\value{
a data.frame
}
\description{
Function used to find the efficient frontier by using the modern portofilio theory
}

