% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effcientFrontierIpop.R
\name{eff.frontier_ipop}
\alias{eff.frontier_ipop}
\title{get the efficient frontier, no matter the covariance matrix is postive semidefinite or not}
\usage{
eff.frontier_ipop(
  returns,
  covMat = NULL,
  short = "yes",
  risk.premium.up = 20,
  thread = 3,
  maxiter = 40
)
}
\arguments{
\item{covMat}{The predefined covariance matrix}

\item{short}{whether short selling is prohibited or not}

\item{thread}{the multi threads number}

\item{maxiter}{Maximum number of iterations}

\item{retruns}{The return matrix of all the assets,T*N dimension}

\item{risk.premim}{the maximum risk aversion parameter}

\item{risk.increment}{the incremental of the risk aversion}
}
\value{
a data.frame include the weight, variance, return and shape ratio
}
\description{
get the efficient frontier, no matter the covariance matrix is postive semidefinite or not
}
