% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/F_DCCA.R
\name{F_DCCA}
\alias{F_DCCA}
\title{Function used to calculate the DCCA fluctuation}
\usage{
F_DCCA(x, y, n, overlap = 0.99999999, q = 2)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{y}{a numeric vector which convert from a time series}

\item{n}{the time scale of the detrended operation}

\item{overlap}{the overlap of the boxes,default 0.99999999}

\item{q}{the order of the multifractal}
}
\value{
the the DCCA fluctuation
}
\description{
Function used to calculate the DCCA fluctuation
}

