% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredCorMat.R
\name{filteredCorMat}
\alias{filteredCorMat}
\title{Function used to filter the covariance matrix of a series of return time series}
\usage{
filteredCorMat(returns)
}
\arguments{
\item{returns}{the return matrix}
}
\value{
a list with four matrices: original,market,structure and random
}
\description{
Function used to filter the covariance matrix of a series of return time series
}
\references{
Random matrix approach to cross correlations in financial data
}
