% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cycles.R
\name{find_cycles}
\alias{find_cycles}
\title{Function used to find the local minimum and local maximum}
\usage{
find_cycles(x)
}
\arguments{
\item{x}{a time series}
}
\value{
two numeric values, the cycle length of local maxima and minimia
}
\description{
Function used to find the local minimum and local maximum
}
