% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGiantComponent.R
\name{findGiantComponent}
\alias{findGiantComponent}
\title{Find the giant component}
\usage{
findGiantComponent(GRAPH, returnGRAPH = FALSE, normalized = TRUE)
}
\arguments{
\item{GRAPH}{a igraph object}

\item{returnGRAPH}{return graph or not}

\item{normalized}{a logical, normalize the vertex number or not}
}
\value{
the size of the giant component
}
\description{
Find the giant component
}
