% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridMeasure.R
\name{hybirdMeasure}
\alias{hybirdMeasure}
\title{Function calculate hybird measure for centrality of graph}
\usage{
hybirdMeasure(GRAPH_W)
}
\arguments{
\item{GRAPH_W}{an weighted igraph object 
return a list or a data.frame}
}
\description{
Function calculate hybird measure for centrality of graph
}
