% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/initializePars.R
\name{initializePars}
\alias{initializePars}
\title{Function used to initialize some system parameters}
\usage{
initializePars()
}
\description{
Function used to initialize some system parameters
}

