% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_MatrixDistance.R
\name{JS_MatrixDistance}
\alias{JS_MatrixDistance}
\title{Function used to calculate the Jensen-Shannon between two matrices, i.e, the distance between eigenvalue distribution}
\usage{
JS_MatrixDistance(mat1, mat2, n = 512, bw = 0.01, thread = 3)
}
\arguments{
\item{mat1}{a matrix}

\item{mat2}{a matrix}

\item{n}{the number of equally spaced points at which the density is to be estimated}

\item{bw}{the bandwidth of the density function}

\item{thread}{the multi threads}
}
\value{
a numeric value between 0 and 1
}
\description{
Function used to calculate the Jensen-Shannon between two matrices, i.e, the distance between eigenvalue distribution
}
