% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_community.R
\name{layout_community}
\alias{layout_community}
\title{Function used to put each community as the same layout}
\usage{
layout_community(graph, community_algorithm, layout_desired)
}
\arguments{
\item{graph}{A igraph object}

\item{algorithm}{A character name for the community detection method}

\item{layout}{A character name for the layout function you use}
}
\value{
A data frame with four columns, the vertex label, the membership of each node, two columns of the coordinates
}
\description{
Function used to put each community as the same layout
}
