% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dcca.R
\name{plot.dcca}
\alias{plot.dcca}
\title{plot the fluctuation function for DCCA}
\usage{
\method{plot}{dcca}(x, xlab = "s", ylab = "fq", ylim = NULL, tck = 0.02)
}
\arguments{
\item{x}{the fluctuation value}

\item{xlab}{the x label}

\item{ylab}{the y label}

\item{ylim}{the y coordinate range}

\item{tck}{the plot tck value for the paramter option}
}
\value{
invisible
}
\description{
plot the fluctuation function for DCCA
}
