% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_data.r
\name{polish_data}
\alias{polish_data}
\title{polish data from the uncomtrade database}
\usage{
polish_data(data, merge = FALSE, weight_method = c("mean", "max",
  "min"))
}
\arguments{
\item{data}{a tidyverse like tibble}

\item{merge}{merge HK and Maco to mainland China or not}
}
\value{
a tidyverse like tibble
}
\description{
polish data from the uncomtrade database
}
