% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_DCCA.R
\name{rho_DCCA}
\alias{rho_DCCA}
\title{Function used to calculate the DCCA coefficient}
\usage{
rho_DCCA(
  x,
  y,
  nVec = NULL,
  sampleNum = NULL,
  qVec,
  detrendOrder = 3,
  thread = 3,
  sampleMethod = 2,
  lengthRatio = 0.05
)
}
\arguments{
\item{x}{a numeric vector which convert from a time series}

\item{y}{a numeric vector which convert from a time series}

\item{nVec}{the time scale of the detrended operation}

\item{sampleNum}{the number of sample}

\item{qVec}{the fluctuation order}

\item{detrendOrder}{the order of the polynomial fit}

\item{thread}{number of the multithread}

\item{sampleMethod}{1 means the determined sample number, other values means the nonoverlap method}

\item{lengthRatio}{the ratio of largest detrending scale over the time series length}
}
\value{
the DCCA coefficient ranging from -1 to 1, a dataframe
}
\description{
Function used to calculate the DCCA coefficient
}
