% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_transimission.r
\name{risk_transmission}
\alias{risk_transmission}
\title{modeling the risk transmission of the trade network}
\usage{
risk_transmission(
  net,
  mode = c("Node_driven", "Edge_driven"),
  n_id = NULL,
  e_id = NULL,
  alpha = 0.1,
  return_net = FALSE
)
}
\arguments{
\item{net}{a tidygraph object}

\item{mode}{two different supply shock scenario}

\item{n_id}{the initial shocked node id}

\item{e_id}{the initial shocked edge id}

\item{alpha}{the threshold above which the infected nodes during the transmission process will change into active state}

\item{return_net}{whether return the final network structure or not}
}
\value{
the active node size for each iteration
}
\description{
modeling the risk transmission of the trade network
}
