% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spilloverNet.r
\name{spilloverNet}
\alias{spilloverNet}
\title{generate the spillover network for a return dataframe}
\usage{
spilloverNet(dat, parallel = TRUE, n_cores = 3)
}
\description{
generate the spillover network for a return dataframe
}
