% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/triplePredict.R
\name{triplePredict}
\alias{triplePredict}
\title{Compute the triplets to evaluate the complexity of the fractal spectrum}
\usage{
triplePredict(h_q, qVec, tol = 1e-06)
}
\arguments{
\item{h_q}{the generalized hurst exponent}

\item{qVec}{the order of q}

\item{tol}{the tolerence of splinfun used to determine the x_min and x_max in the fractal spectrum}
}
\value{
A vector with three elements
}
\description{
Compute the triplets to evaluate the complexity of the fractal spectrum
}

